requirements.txt
pyTelegramBotAPI
python-dotenv
Flask
gunicorn
Python-IO
psycopg2-binary